add 0.2.0 all context

+++ Old8Lang/CslyMake/OldExpression/OldExpr/OldValue/OldItem.cs
@@ -25,7 +25,7 @@ public class OldItem : OldValue
         {
             var intResult = result as OldInt;
             var alist = a as OldList;
-            return alist.Value[intResult.Value];
+            return alist.Value[(int)intResult.Value];
         }
 
         if (a is OldDir && result is not null)
